+++ spec/spec_helper.rb
@@ -52,3 +52,7 @@ if ENV['TRAVIS']
   # Don't execute tests against "merge" commits
   exit 0 if ENV['TRAVIS_PULL_REQUEST'] != 'false' && ENV['TRAVIS_BRANCH'] == 'master'
 end
+
+def fixture_path(path)
+  Pathname.new(__FILE__).join("../fixtures").expand_path.join(path)
+end